requirements.txt
python-telegram-bot
python-dotenv
streamlit
flask
cryptography